[JSON] JSONElements should be able to be larger than 64Kb (uint16_t)
Although the JSON parsing can be done in segement and thus the parser is capable of working
with a maxSize, per segment, of 65535 bytes (uiny16_t), position determination in the generated
object (indicated by the offset) should exceed the 64Kb boundary. For this reason, changed the
offset handling from a uint16_t to a uint32_t.

+++ Source/websocket/WebSerializer.h
@@ -390,7 +390,7 @@ namespace Web {
     private:
         mutable uint32_t _lastPosition;
         mutable string _body;
-        uint16_t _offset;
+        uint32_t _offset;
     };
 
     template <typename JSONOBJECT, typename HASHALGORITHM>
